// projects/angular/src/utils/popover/enums/alignment.enum.ts
/*
 * Copyright (c) 2016-2024 Broadcom. All Rights Reserved.
 * The term "Broadcom" refers to Broadcom Inc. and/or its subsidiaries.
 * This software is released under MIT license.
 * The full license information can be found in LICENSE in the root directory of this project.
 */

export enum ClrAlignment {
  START = 0,
  CENTER = 0.5,
  END = 1,
}